changed some file names
